added checks for files ending in fastq
