+++ pipelines/skylab/optimus/scripts/parse_terra_tsv.py
@@ -13,7 +13,7 @@ Outputs:
     Sample set TSV to be uploaded to terra. The samples will be grouped by sample_id.
 """
 
-def create_output_files(input_file,output_file,output_set):
+def create_output_files(input_file,output_file,output_set,output_entity):
     """
     Args:
         input_file: tsv file from HCA
@@ -27,24 +27,24 @@ def create_output_files(input_file,output_file,output_set):
 
 #TBD: move this to a function and call on each row of df change. Note: change index 0 to get other participants
     # for each fastq read, create a lane
-    n_lanes = r1_fastq.shape[1] #number of fastq reads
+    n_lanes = r1_fastq.shape[1] - r1_fastq.isnull().sum(axis=1) #number of fastq reads
     n_participants = r1_fastq.shape[0] #number of participants
     column_names = ['entity:participant_lane_id', 'input_id', 'input_name','input_id_metadata_field','input_name_metadata_field', 'r1_fastq','r2_fastq', 'i1_fastq']
     participant_df = pd.DataFrame(columns = column_names)
 
     for j in range(n_participants):
         a  = []
-        for i in range(n_lanes):
+        for i in range(n_lanes[j]):
             a.append("participant_"+str(j)+"_lane_"+str(i)+"_"+str(df.sequencing_process__provenance__document_id[j])+"_id")
     
         lane_id = pd.DataFrame({"entity:participant_lane_id":a})
         lane_fastq_r1 = pd.DataFrame({"fastq1":r1_fastq.iloc[j].to_numpy()})
         lane_fastq_r2 = pd.DataFrame({"fastq2":r2_fastq.iloc[j].to_numpy()})
         lane_fastq_i1 = pd.DataFrame({"fastqi":i1_fastq.iloc[j].to_numpy()})
-        input_id = pd.DataFrame({"input_id":np.repeat(df.sequencing_process__provenance__document_id[j],n_lanes)})
-        input_id_metadata_field = pd.DataFrame({"input_id_metadata_field":np.repeat("sequencing_process.provenance.document_id",n_lanes)})
-        input_name = pd.DataFrame({"input_name_metadata_field":np.repeat(df.sequencing_input__biomaterial_core__biomaterial_id[j],n_lanes)})
-        input_name_metadata_field = pd.DataFrame({"entity:lane_id":np.repeat("sequencing_input.biomaterial_core.biomaterial_id",n_lanes)})
+        input_id = pd.DataFrame({"input_id":np.repeat(df.sequencing_process__provenance__document_id[j],n_lanes[j])})
+        input_id_metadata_field = pd.DataFrame({"input_id_metadata_field":np.repeat("sequencing_process.provenance.document_id",n_lanes[j])})
+        input_name = pd.DataFrame({"input_name_metadata_field":np.repeat(df.sequencing_input__biomaterial_core__biomaterial_id[j],n_lanes[j])})
+        input_name_metadata_field = pd.DataFrame({"entity:lane_id":np.repeat("sequencing_input.biomaterial_core.biomaterial_id",n_lanes[j])})
 
         column_names = ['entity:participant_lane_id', 'input_id', 'input_name','input_id_metadata_field','input_name_metadata_field', 'r1_fastq','r2_fastq', 'i1_fastq']
 
@@ -60,12 +60,19 @@ def create_output_files(input_file,output_file,output_set):
                    axis=1)
         lane_df.columns = column_names
         participant_df = participant_df.append(lane_df)      
-    participant_df.to_csv(output_file,sep="\t",index=None)
+    participant_lane_df = participant_df.dropna()
+
+    participant_lane_df.to_csv(output_file,sep="\t",index=None)
     #print(out_df.shape,out_df.columns,out_df.r2_fastq)
 
     particpant_set_df = participant_df[['input_id','entity:participant_lane_id']]
     particpant_set_df.columns = ['membership:participant_lane_set_id', 'participant_lane']
     particpant_set_df.to_csv(output_set,sep="\t",index=None)
+
+    temp = df[['sequencing_process__provenance__document_id','sequencing_input__biomaterial_core__biomaterial_id']]
+    temp.columns = ['entity:participant_lane_set_id','input_name']
+    temp.to_csv(output_entity,sep="\t",index=None)
+
 
 def main():
     description = """This script converts the tsv file from HCA to data table to be used in terra.
@@ -91,10 +98,16 @@ def main():
         required=True,
         help="Sample set TSV to be uploaded to terra"
     )
+    parser.add_argument(
+        "--output_entity",
+        dest="output_entity",
+        required=True,
+        help="Additional columns for the sample set TSV to be uploaded to terra"
+    )
     args = parser.parse_args()
     #print(args.output_file)
 
-    create_output_files(args.input_file, args.output_file, args.output_set)
+    create_output_files(args.input_file, args.output_file, args.output_set,args.output_entity)
 
 
 if __name__ == "__main__":
