+++ tests/skylab/smartseq2_single_sample/pr/test_inputs.json
@@ -8,8 +8,8 @@
   "TestSmartSeq2SingleCellPR.gene_ref_flat": "gs://hca-dcp-mint-test-data/reference/GRCh38_Gencode/GRCh38_gencode.v27.refFlat.txt",
   "TestSmartSeq2SingleCellPR.hisat2_ref_name": "genome_snp_tran",
   "TestSmartSeq2SingleCellPR.stranded":"NONE",
-  "TestSmartSeq2SingleCellPR.fastq1":"gs://hca-dcp-mint-test-data/smartseq2_single_sample/patel_ap/SRR1294925_1.fastq.gz",
-  "TestSmartSeq2SingleCellPR.fastq2":"gs://hca-dcp-mint-test-data/smartseq2_single_sample/patel_ap/SRR1294925_2.fastq.gz",
+  "TestSmartSeq2SingleCellPR.fastq1":"gs://hca-dcp-mint-test-data/smartseq2_single_sample/patel_ap/SRR1294925_1",
+  "TestSmartSeq2SingleCellPR.fastq2":"gs://hca-dcp-mint-test-data/smartseq2_single_sample/patel_ap/SRR1294925_2",
   "TestSmartSeq2SingleCellPR.input_id":"SRR1294925",
   "TestSmartSeq2SingleCellPR.output_name":"SRR1294925",
   "TestSmartSeq2SingleCellPR.expected_counts_hash": "135a3fbb959583db17713dc8b9d7fe33",
